pyproject.toml
[project]
name = "travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.12.1",
    "fastapi[standard]>=0.135.2",
    "fastmcp>=3.1.1",
    "pydantic-ai>=1.73.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
]

[tool.pytest.ini_options]
# Automatically run all 'async def' tests without @pytest.mark.asyncio
asyncio_mode = "auto"
